scraper/requirements.txt
requests>=2.25.1
python-dotenv>=0.19.0
tenacity>=8.0.0
pathlib2>=2.3.5; python_version < "3.4"
Pillow>=9.0.0 